Client-Server Application in C++
In this project I implemented a TCP connection between a server and multiple clients. The server received information on a specific topic from UDP clients. The TCP clients could (un)subscribe to a topic and then they neeed to (not)receive every message that was sent on that topic. Also they could disconnect and the rejoin the connection.