Assembly X86 - decrypting messages
The purpose of this project was to apply the notions about stack and memory; the whole project was implemented in assembly language. Algorithms such as one time pad, base32, brute force on xor with one byte key, Vigenere chipher have been implemented. It was an interesting project because I learned the back notions of the function call in C.