Modelling dense graph in C++
Within this project I had to model the map of a country seen as a graph. The project was quite challenging because there was a time limit, and the entrance tests had many nodes and edges; Adjacent lists and more advanced algorithms for the latest tasks, for example Kosaraju / Tarhan, had to be used.