ER Simulator
In this project I had to simulate a hospital with different entities: doctors, assistants, technicians, patients. There were several stages: nurses saw the patients, doctors examined the patients and if necessary technicians examined the patients. The aim was to simulate the real activities of a hospital and to use the Observer pattern design.