MNIST and Audiobooks - ML problems
These are the first problems which I made that involve deep learning. In the first one I used the MNIST dataset provided from TensorFlow datasests which contained images with digits from 0 to 9. The scope was to make a ML algorithm that could classify the images that it received as inputs. The approch was a neuronal network solution wih five hidden layers and the use of adam optimizer for an loss entropy function; the result was a 97% accuracy on the test samples.
In the second problem I was provided with a dataset from an Audiobook store and the goal was to predict if a client will come back to the store based on the information gathered. I used a neuronal network with  six hidden layers and the adam optimizer for the loss entropy function. It resulted in a 83% accuracy.