Optical Recognition on handwritten numbers
In this project I had to apply the knowledge learned in Data Structures Course within a more complex program. We received a code skeleton and had to implement certain functions. The project was interesting because I analyzed code written by other people and added functionalities to it; we had to work in teams, which was very nice.
 We had to implement a variant of the Random Forest artificial intelligence algorithm, which used decision trees. The program received at the input the image of a digit from 0 to 9, represented as a vector of 28 * 28 elements) and returned with an accuracy of about 90% the respective digit.