Implementing Sliding Window Protocol in C
In this project I implemented a sliding window protocol in C lanuage. We were provided with a link emulator which emulated a direct link between the sender and the receiver. It was a challenging project because i have to consider some cases: the case when the link doesn't loss and don't corrupt the data, the case when the link loss some data, but doesn't corrut it, the case when both can happen and the final one was the one in which the frames could arrive and the sender in other order.